\name{memuse-package}
\alias{memuse-package}
\docType{package}
\title{Core memuse Classes and Methods}
\description{
  The package gives the user the size (memory usage) of an in-core, 
  dense matrix. The output is an S4 class object that can be manipulated 
  to be presented in different ways (different units, short/long 
  versions of those units, etc.). For more information, see the package 
  vignette and the manual.
}
\details{
\tabular{ll}{
  Package: \tab memuse\cr
  Type: \tab Package\cr
  License: \tab GPL\cr
  LazyLoad: \tab yes\cr
}
}
\references{
  \url{http://wrathematics.github.io}
}
\author{
  Drew Schmidt
}
\keyword{Package}
