% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.r
\name{approx.size}
\alias{approx.size}
\alias{hr}
\title{approx.size}
\usage{
approx.size(x, unit.names = "long", ..., digits = 1)

hr(x)
}
\arguments{
\item{x}{A number.}

\item{unit.names}{"long", "short", or "comma"; determines wheter the output
reads like "10 million", "10m", or "10,000,000", respectively.}

\item{...}{Additional arguments}

\item{digits}{The number of decimal digits to retain.}
}
\value{
An object of class \code{approx}.
}
\description{
A simple printing system to make numbers readable.

\code{hr()}, short for "human readable", is an alias for \code{approx.size()}
with \code{unit.names="comma"}.
}
\details{
Approximate size of an integer; a poor man's exponential notation.
}
\examples{
\dontrun{
library(memuse)

approx.size(12345678)
approx.size(12345678, unit.names="comma")
}

}
\seealso{
\code{\link{howmany}}
}
