% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meminfo.r
\name{cachelinesize}
\alias{Sys.cachelinesize}
\alias{cachelinesize}
\title{Cache Sizes and Linesize}
\usage{
Sys.cachelinesize()
}
\value{
Returns a list, whose values are platform dependent.
}
\description{
Shows the size of the cache line.
}
\details{
\code{Sys.cachelinesize()} will return the cache linesize.  It's almost
certainly 32 or 64 bytes.
}
\examples{
\dontrun{
library(memuse)

Sys.cachelinesize()
}
}
\seealso{
\code{\link{meminfo}}
}

