% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meminfo.r
\name{cachesize}
\alias{Sys.cachesize}
\alias{cachesize}
\title{Cache Sizes}
\usage{
Sys.cachesize()
}
\value{
Returns a list, whose values are platform dependent.
}
\description{
Shows the sizes of the CPU caches.
}
\details{
\code{Sys.cachesize()} will check the various levels of cache and return all
available cache information in a list.  If you don't have some kind of
level-1 cache, then it will return an error.  If you have some kind of
future space computer with more than 3 levels of cache, levels higher than 3
will not be displayed.
}
\examples{
\dontrun{
library(memuse)

Sys.cachesize()
}

}
\seealso{
\code{\link{meminfo}}
}

