% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparators.r
\name{Comparators}
\alias{Comparators}
\alias{==,memuse,memuse-method}
\alias{<,memuse,memuse-method}
\alias{<=,memuse,memuse-method}
\alias{>,memuse,memuse-method}
\alias{>=,memuse,memuse-method}
\alias{!=,memuse,memuse-method}
\alias{==,memuse,numeric-method}
\alias{<,memuse,numeric-method}
\alias{<=,memuse,numeric-method}
\alias{>,memuse,numeric-method}
\alias{>=,memuse,numeric-method}
\alias{!=,memuse,numeric-method}
\alias{==,numeric,memuse-method}
\alias{<,numeric,memuse-method}
\alias{<=,numeric,memuse-method}
\alias{>,numeric,memuse-method}
\alias{>=,numeric,memuse-method}
\alias{!=,numeric,memuse-method}
\title{Comparators}
\usage{
\S4method{==}{memuse,memuse}(e1, e2)

\S4method{<}{memuse,memuse}(e1, e2)

\S4method{<=}{memuse,memuse}(e1, e2)

\S4method{>}{memuse,memuse}(e1, e2)

\S4method{>=}{memuse,memuse}(e1, e2)

\S4method{!=}{memuse,memuse}(e1, e2)

\S4method{==}{memuse,numeric}(e1, e2)

\S4method{<}{memuse,numeric}(e1, e2)

\S4method{<=}{memuse,numeric}(e1, e2)

\S4method{>}{memuse,numeric}(e1, e2)

\S4method{>=}{memuse,numeric}(e1, e2)

\S4method{!=}{memuse,numeric}(e1, e2)

\S4method{==}{numeric,memuse}(e1, e2)

\S4method{<}{numeric,memuse}(e1, e2)

\S4method{<=}{numeric,memuse}(e1, e2)

\S4method{>}{numeric,memuse}(e1, e2)

\S4method{>=}{numeric,memuse}(e1, e2)

\S4method{!=}{numeric,memuse}(e1, e2)
}
\arguments{
\item{e1, e2}{\code{memuse}, \code{numeric}, or \code{object_size} objects.}
}
\value{
Returns a \code{memuse} class object.
}
\description{
Binary comparators for \code{memuse} objects.
}
\details{
Comparisons to numeric values are done at the byte level.
}
\examples{
\dontrun{
x <- mu(2000)
y <- mu(3000)

x < y
x <= y
x > y
x >= y
x == y
x != y
}

}
\seealso{
\code{ \link{Constructor}, \link{memuse-class} }
}
\keyword{Methods}
