% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr.r
\name{hr}
\alias{hr}
\title{hr}
\usage{
hr(x, names = "comma", digits = 1)
}
\arguments{
\item{x}{A number.}

\item{names}{"long", "short", or "comma"; determines wheter the output
reads like "10 million", "10m", or "10,000,000", respectively.}

\item{digits}{The number of decimal digits to retain.}
}
\value{
An object of class \code{humanreadable}.
}
\description{
A poor man's exponential notation.
}
\details{
Approximate size of an integer.  Very useful when dealing with potentially
large values, such as those from \code{howmany()}.
}
\examples{
\dontrun{
library(memuse)

hr(12345678)
hr(12345678, "long")
}

}
\seealso{
\code{\link{howmany}}
}
