% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meminfo.r
\name{meminfo}
\alias{Sys.meminfo}
\alias{meminfo}
\title{meminfo}
\usage{
Sys.meminfo(compact.free = TRUE)
}
\arguments{
\item{compact.free}{logical; determines whether various free memory values
should be combined into a single value. See details section for more
information.}
}
\value{
Returns a list, whose values are platform dependent in addition to
being modified by input arguments.
}
\description{
Platform memory information.
}
\details{
\code{Sys.meminfo()} returns some basic memory values, such as total ram,
free ram, and ram used for buffers/cache (when applicable).

All of the C-level source code for these methods (in src/meminfo of the root
directory of the memuse source tree) is licensed under the permissive
2-Clause BSD license.
}
\examples{
\dontrun{
library(memuse)

Sys.meminfo()
}

}
\seealso{
\code{\link{procmem}}
}

