% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memuse-package.R
\docType{package}
\name{memuse-package}
\alias{memuse-package}
\title{Core memuse Classes and Methods}
\description{
Originally an amusing fancy printing system, this package contains that, as
well as some functionality for showing.  As of version 2.0, the package
contains some helpful utilities for accessing hardware ram information, ram
usage of the current R process, as well as some other niceties.
}
\details{
If you do a lot of benchmarking, this is the package you've been waiting
for.  Maybe.  I mean, I don't want to be too pushy about it.
}
\references{
Project home page: \url{https://shinra-dev.github.io}
}
\author{
Drew Schmidt
}
\keyword{Package}
