% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meminfo.r
\name{post-system-time}
\alias{post-system-time}
\alias{post.system.time}
\title{post.system.time}
\usage{
post.system.time()
}
\value{
A proc_time object (same as \code{system.time()}).
}
\description{
Returns the user, kernel, and wallclock time that the current
process has been active.  Like a kind of post-hoc \code{system.time()}.
}
\details{
All of the C-level source code for these methods (in src/meminfo of the root
directory of the memuse source tree) is licensed under the permissive
2-Clause BSD license.
}
\examples{
\dontrun{
library(memuse)

post.system.time()
}
}

