% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr.r
\name{print-hr}
\alias{print-hr}
\alias{print.humanreadable}
\title{Print \code{humanreadable} objects}
\usage{
\method{print}{humanreadable}(x, ...)
}
\arguments{
\item{x}{\code{approx} object}

\item{...}{unused}
}
\description{
Printing for \code{hr()}
}
