% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\docType{methods}
\name{print-memuse}
\alias{print-memuse}
\alias{print,memuse-method}
\alias{show,memuse-method}
\title{Printing}
\usage{
\S4method{print}{memuse}(x, ..., unit = mu.unit(x), prefix = mu.prefix(x),
  names = mu.names(x), digits = 3)

\S4method{show}{memuse}(object)
}
\arguments{
\item{x, object}{\code{memuse} class object}

\item{...}{extra arguments}

\item{unit}{the unit to be used in printing; defaults to \code{x}'s unit}

\item{prefix}{the unit prefix to be used in printing; defaults to
\code{x}'s prefix}

\item{names}{the unit names (short or long) to be used in printing;
defaults to \code{x}'s names}

\item{digits}{the number of decimal digits to print; default is 3}
}
\description{
Print and show methods for \code{memuse} class objects.
}
\examples{
\dontrun{
x <- mu(1e6)

print(x)
x # same as show(x)
}

}
\seealso{
\code{\ link{Constructor} \link{memuse-class} }
}
\keyword{Methods}
