% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timings.r
\name{print-readabletime}
\alias{print-readabletime}
\alias{print.readabletime}
\title{Print readabletime objects.}
\usage{
\method{print}{readabletime}(x, ..., digits = 3)
}
\arguments{
\item{x}{An readabletime object.}

\item{...}{Additional arguments (ignored).}

\item{digits}{The (max) number of digits to display.}
}
\description{
Print readabletime objects.
}

