% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.readabletime}
\alias{print.readabletime}
\title{Print readabletime objects.}
\usage{
print.readabletime(x, ..., digits = 3)
}
\arguments{
\item{x}{A count, in seconds, to be printed.}

\item{...}{Extra arguments (currently ignored).}

\item{digits}{The number of decimal digits to print.}
}
\description{
Print readabletime objects.
}

