% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sum.r
\docType{methods}
\name{Reductions}
\alias{Reductions}
\alias{sum,memuse-method}
\title{memuse Arithmetic}
\usage{
\S4method{sum}{memuse}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A \code{memuse} object.}

\item{...}{Additional arguments}

\item{na.rm}{Whether \code{NA}'s should be ignored.}
}
\value{
Returns a \code{memuse} class object.
}
\description{
Binary arithmetic operations for \code{memuse} objects.
}
\details{
Simple arithmetic reductions.
}
\seealso{
\code{ \link{Constructor} \link{memuse-class} }
}
\keyword{Methods}

