% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meminfo.r
\name{swapinfo}
\alias{Sys.pageinfo}
\alias{Sys.swapinfo}
\alias{swapinfo}
\title{swapinfo}
\usage{
Sys.swapinfo()

Sys.pageinfo()
}
\value{
Returns a list, whose values are platform dependent in addition to
being modified by input arguments.
}
\description{
Platform swap information.
}
\details{
\code{Sys.swapinfo()} returns basic swap/page (virtual memory) information.
\code{Sys.pageinfo()} is identical to \code{swapinfo()} in every way but
name (provided for Windows users who may be more comfortable/familiar with
the 'pagefile' naming convention).

All of the C-level source code for these methods (in src/meminfo of the root
directory of the memuse source tree) is licensed under the permissive
2-Clause BSD license.
}
\examples{
\dontrun{
library(memuse)

Sys.swapinfo()
}
}
\seealso{
\code{\link{procmem}}
}
\keyword{Methods}

