% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meminfo.r
\name{systime}
\alias{Sys.systime}
\alias{systime}
\title{systime}
\usage{
Sys.systime()
}
\value{
A readabletime object.
}
\description{
Returns the amount of time, in human-readable units, the
current process has been in kernel mode.
}
\details{
All of the C-level source code for these methods (in src/meminfo of the root
directory of the memuse source tree) is licensed under the permissive
2-Clause BSD license.
}
\examples{
\dontrun{
library(memuse)

Sys.systime()
}
}

