% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meminfo.r
\name{uptime}
\alias{Sys.uptime}
\alias{uptime}
\title{uptime}
\usage{
Sys.uptime()
}
\value{
A readabletime object.
}
\description{
Returns the amount of time, in human-readable units, since
the system was booted.
}
\details{
All of the C-level source code for these methods (in src/meminfo of the root
directory of the memuse source tree) is licensed under the permissive
2-Clause BSD license.
}
\examples{
\dontrun{
library(memuse)

Sys.uptime()
}
}

