% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timings.r
\name{wc.time}
\alias{wc.time}
\title{Readable Timings}
\usage{
wc.time(expr, gcFirst = TRUE)
}
\arguments{
\item{expr}{A valid R expression to be timed.}

\item{gcFirst}{logical; determines if garbage collection should be called
before getting process memory usage.}
}
\value{
Returns a \code{timing} object, which dispalys readable times.
}
\description{
Printing timings in their "native" unit.  For example, "10000" seconds
becomes "2.778 hours".
}
\details{
The wall-clock timer, \code{wc.time()}, is a simple wrapper around R's
\code{system.time()} using these readable units.

This uses S3 methods for simplicity, though this makes some higher
abstractions impossible.  Future versions will likely use S4.
}
\examples{
\dontrun{
library(memuse)

wc.time(rnorm(1e6))
}
}
\keyword{Methods}

