\name{memuse.internal}

% Virtual classes and methods
\alias{approx-class}
\alias{print,approx-method}
\alias{show,approx-method}

\alias{object.size}
\alias{object.size-method}
\alias{object.size,ANY-method}


% Internal data structures
\alias{.units}
\alias{.numbers}
\alias{.time}
\alias{.units_bits}


\title{ Internal Functions }
\description{ Internal functions }
\keyword{internal}

