\name{memuse-package}
\alias{memuse-package}
\docType{package}
\title{Core memuse Classes and Methods}
\description{
  The package gives the user the size (memory usage) of an unallocated, dense, 
  in-core numeric object (matrix/vector). The output is an S4 class object 
  that can be manipulated to be presented in different ways (different units, 
  short/long versions of those units, etc.). Details about these various 
  options are outlined in the rest of this manual, as well as the package 
  vignette.
}
\details{
\tabular{ll}{
  Package: \tab memuse\cr
  Type: \tab Package\cr
  License: \tab GPL\cr
  LazyLoad: \tab yes\cr
}
}
\references{
  Project home page:
  \url{http://wrathematics.github.io}

  CRAN listing:
  \url{http://cran.r-project.org/web/packages/memuse/index.html}
}
\author{
  Drew Schmidt
}
\keyword{Package}
