\name{pbdBASE-package}
\alias{pbdBASE-package}
\docType{package}
\title{Core pbd Classes and Methods}
\description{
  A package contains the basic methods for dealing with 
  distributed data types, as well as the data types themselves.
}
\details{
\tabular{ll}{
  Package: \tab pbdBASE\cr
  Type: \tab Package\cr
  License: \tab GPL\cr
  LazyLoad: \tab yes\cr
}
  This package requires an MPI library (OpenMPI, MPICH2, or LAM/MPI).
}
\references{
  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
}
\author{
  Drew Schmidt \email{schmidt AT math.utk.edu}, Wei-Chen Chen,
  George Ostrouchov, and Pragneshkumar Patel.
}
\keyword{Package}
