% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr.r
\name{hr}
\alias{hr}
\title{hr}
\usage{
hr(x, unit.names = "comma", digits = 1)
}
\arguments{
\item{x}{A number.}

\item{unit.names}{"long", "short", or "comma"; determines wheter the output
reads like "10 million", "10m", or "10,000,000", respectively.}

\item{digits}{The number of decimal digits to retain.}
}
\value{
An object of class \code{approx}.
}
\description{
A simple printing system to make numbers readable.

\code{hr()}, short for "human readable", is an alias for \code{approx.size()}
with \code{unit.names="comma"}.
}
\details{
Approximate size of an integer; a poor man's exponential notation.
}
\examples{
\dontrun{
library(memuse)

hr(12345678)
hr(12345678, "long")
}

}
\seealso{
\code{\link{howmany}}
}
