% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysinfo.r
\name{filesize}
\alias{filesize}
\alias{Sys.filesize}
\alias{Sys.dirsize}
\title{filesize}
\usage{
Sys.filesize(filename)

Sys.dirsize(dirname)
}
\arguments{
\item{filename, dirname}{Location of the file/directory (as a string).}
}
\value{
A memuse class object.
}
\description{
Returns size of a file as reported by the file system (the
file is not scanned).
}
\details{
All of the C-level source code for these methods (in src/meminfo of the root
directory of the memuse source tree) is licensed under the permissive
2-Clause BSD license.
}
\examples{
\dontrun{
library(memuse)

x <- rnorm(1e5)
memuse(x) ### size in ram

tmp <- tempfile()
saveRDS(x, file=tmp)
Sys.filesize(tmp) ### size on disk
unlink(tmp)
}

}
