% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/00_classes.r
\docType{class}
\name{memuse-class}
\alias{memuse-class}
\title{Class memuse}
\description{
Memory usage class object.
}
\section{Slots}{

\describe{
\item{\code{size}}{TODO}

\item{\code{unit}}{TODO}

\item{\code{unit.prefix}}{TODO}

\item{\code{unit.names}}{TODO

\code{memuse(size = ..., unit = ..., unit.prefix = ..., unit.names = ...)}

\code{mu(size = ..., unit = ..., unit.prefix = ..., unit.names = ...)}}
}}
\section{Creating Objects}{
 \code{new('memuse', size = ..., unit = ...,
unit.prefix = ..., unit.names = ...)}
}
\seealso{
\code{ \link{Control} \link{Constructor} }
}
\keyword{Classes}

