% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.r
\docType{class}
\name{memuse-class}
\alias{memuse-class}
\title{Class memuse}
\description{
Memory usage class object.
}
\section{Slots}{

\describe{
\item{\code{size}}{The actual size in some memuse units.}

\item{\code{unit}}{The mem unit (e.g., byte, kilobyte, etc.)}

\item{\code{unit.prefix}}{IEC or SI units}

\item{\code{unit.names}}{short (e.g., kb) or long (e.g., kilobyte)}
}}
\seealso{
\code{ \link{Control} \link{Constructor} }
}
\keyword{Classes}

