% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{memuse-package}
\alias{memuse-package}
\title{Core memuse Classes and Methods}
\description{
Originally an amusing fancy printing system, this package contains that, as
well as some functionality for showing.  As of version 2.0, the package
contains some helpful utilities for accessing hardware ram information, ram
usage of the current R process, as well as some other niceties.
}
\details{
If you do a lot of benchmarking, this is the package you've been waiting
for.  Maybe.  I mean, I don't want to be too pushy about it.

\tabular{ll}{
   Package: \tab memuse \cr
   Type: \tab Package \cr
   License: \tab GPL (>= 2) \cr
   LazyData: \tab yes \cr
   LazyLoad: \tab yes \cr
   NeedsCompilation: \tab yes
}
}
\author{
Drew Schmidt
}
\references{
Project home page: \url{http://github.com/wrathematics/memuse}

CRAN listing: \url{http://cran.r-project.org/web/packages/memuse/index.html}
}
\keyword{Package}

