% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.r
\name{print-approx}
\alias{print-approx}
\alias{print.approx}
\title{Print \code{approx} objects}
\usage{
\method{print}{approx}(x, ...)
}
\arguments{
\item{x}{\code{approx} object}

\item{...}{unused}
}
\description{
Printing for \code{approx.size()}
}

