% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/approx.r
\name{print.approx}
\alias{print.approx}
\alias{show.approx}
\title{Prints approx objects.}
\usage{
print.approx(x, ...)

show.approx(object)
}
\arguments{
\item{x,object}{An \code{approx} object.}

\item{...}{Extra arguments (ignored).}
}
\description{
Prints approx objects.
}

