% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/approx.r
\name{print-approx}
\alias{print-approx}
\alias{print.approx}
\title{Print approx objects.}
\usage{
\method{print}{approx}(x, ...)
}
\arguments{
\item{x}{An approx object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print approx objects.
}

