% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timings.r
\name{readable.time}
\alias{readable.time}
\title{Print times in a readable way.}
\usage{
readable.time(time, unit = "seconds")
}
\arguments{
\item{time}{A time count.}

\item{unit}{The unit of time (e.g. "seconds").}
}
\description{
Print times in a readable way.
}
\examples{
\dontrun{
library(memuse)

readable.time(10000, unit="seconds")
readable.time(10000, unit="minutes")
}
}

