% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\docType{methods}
\name{Replacers}
\alias{Replacers}
\alias{mu.size<-}
\alias{mu.unit<-}
\alias{mu.prefix<-}
\alias{mu.names<-}
\alias{mu.size<-,memuse-method}
\alias{mu.unit<-,memuse-method}
\alias{mu.prefix<-,memuse-method}
\alias{mu.names<-,memuse-method}
\title{Replacers}
\usage{
mu.size(x) <- value

mu.unit(x) <- value

mu.prefix(x) <- value

mu.names(x) <- value

\S4method{mu.size}{memuse}(x) <- value

\S4method{mu.unit}{memuse}(x) <- value

\S4method{mu.prefix}{memuse}(x) <- value

\S4method{mu.names}{memuse}(x) <- value
}
\arguments{
\item{x}{memuse object}

\item{value}{replacement value}
}
\value{
Returns a numeric element in the case of \code{size()}, otherwise a
string is returned.
}
\description{
Replacement methods for slots of objects of class \code{memuse}.
}
\details{
These methods are syntactic sugar for assignment using ordinary S4
accessors.  So for example, \code{size(x) <- 10} is semantically no
different from calling \code{x@size <- 10}

These methods are strict replacement methods; if you need to swap the units
of a \code{memuse} class object, you should probably be using the
\code{\link{Swaps}} methods.  See example below for further details.
}
\examples{
\dontrun{
x <- mu(2000, unit="bytes")
x

mu.size(x) <- 1000
x
}

}
\seealso{
\code{ \link{Accessors} \link{memuse-class} }
}
\keyword{Methods}
