% Generated by roxygen2 (4.0.2): do not edit by hand
\name{usrtime}
\alias{Sys.runtime}
\alias{Sys.usrtime}
\alias{usrtime}
\title{usrtime}
\usage{
Sys.usrtime()

Sys.runtime()
}
\value{
A readabletime object.

A readabletime object.
}
\description{
Returns the amount of time, in human-readable units, the
current process has been in user mode.

Returns the amount of time, in human-readable units, the
current R process has been active (the wall-clock time of the
process).
}
\details{
All of the C-level source code for these methods (in src/meminfo of the root
directory of the memuse source tree) is licensed under the permissive
2-Clause BSD license.

All of the C-level source code for these methods (in src/meminfo of the root
directory of the memuse source tree) is licensed under the permissive
2-Clause BSD license.
}
\examples{
\dontrun{
library(memuse)

Sys.usrtime()
}
\dontrun{
library(memuse)

Sys.runtime()
}
}

